Move Parameters class
