+++ psalm.xml
@@ -46,7 +46,7 @@
 
         <DuplicateClass>
             <errorLevel type="suppress">
-                <file name="src/Bundle/Controller/Parameters.php" />
+                <file name="src/Component/Symfony/Request/Parameters.php" />
             </errorLevel>
         </DuplicateClass>
