+++ src/deprecation.php
@@ -31,6 +31,7 @@ foreach ($deprecatedInterfaces as $oldInterfaceName => $interfaceName) {
 }
 
 $deprecatedClassesWithAliases = [
+    Sylius\Bundle\ResourceBundle\Controller\Parameters::class => Sylius\Component\Resource\Symfony\Request\Parameters::class,
     Sylius\Bundle\ResourceBundle\Controller\ParametersParser::class => Sylius\Component\Resource\Symfony\Request\ParametersParser::class,
     Sylius\Bundle\ResourceBundle\Controller\RequestConfiguration::class => Sylius\Component\Resource\Symfony\Request\RequestConfiguration::class,
